+++ app/Containers/AppSection/Authentication/UI/API/Transformers/TokenTransformer.php
@@ -2,7 +2,7 @@
 
 namespace App\Containers\AppSection\Authentication\UI\API\Transformers;
 
-use App\Containers\AppSection\Authentication\Data\Dto\Token;
+use App\Containers\AppSection\Authentication\Data\DTOs\Token;
 use App\Ship\Parents\Transformers\Transformer as ParentTransformer;
 
 final class TokenTransformer extends ParentTransformer
